Biggest Commit EVER

+++ src/main/java/com/nopay/nopayapi/service/products/CategoryService.java
@@ -1,8 +1,12 @@
 package com.nopay.nopayapi.service.products;
 
 import com.nopay.nopayapi.entity.products.Category;
+import com.nopay.nopayapi.entity.users.Role;
+import com.nopay.nopayapi.entity.users.User;
 import com.nopay.nopayapi.repository.products.CategoryRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -26,12 +30,42 @@ public class CategoryService {
     }
 
     @Transactional
+    public Category findByName(String name) {
+        Category cat = categoryRepository.findByName(name);
+        if (cat == null) {
+            return null;
+        }
+        return cat;
+    }
+
+    @Transactional
     public Category save(Category category) {
+
+        // only admin role can create a category
+        if (SecurityContextHolder.getContext().getAuthentication() == null) {
+            throw new UsernameNotFoundException("User not found");
+        }
+        User admin = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (!admin.getRole().equals(Role.ADMIN)) {
+            throw new UsernameNotFoundException("Only Admin can create a category");
+        }
         return categoryRepository.save(category);
     }
 
     @Transactional
     public void deleteById(Integer id) {
+
+        // only admin role can delete a category
+        if (SecurityContextHolder.getContext().getAuthentication() == null) {
+            throw new UsernameNotFoundException("User not found");
+        }
+        User admin = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (!admin.getRole().equals(Role.ADMIN)) {
+            throw new UsernameNotFoundException("Only Admin can create a category");
+        }
+
         categoryRepository.deleteById(id);
     }
 }
